renames types folder to propTypes for clarity
