+++ src/component/propTypes/OnRemoveTodo.ts
